feat: reorganize middleware into separate packages and update server
- Create middleware packages: cors, logging, recovery, requestid, validation
- Move HTTP middleware from server to respective middleware packages
- Migrate gRPC interceptors to middleware packages
- Update server package to use new middleware packages
- Add backward compatibility wrappers in server package
- Update go.work to include all middleware packages
- Add root README.md with overview of all packages
