+++ middleware/cors/go.mod
@@ -1,0 +1,4 @@
+module github.com/rshelekhov/golib/middleware/cors
+
+go 1.24.2
+
