+++ middleware/logging/go.mod
@@ -1,0 +1,15 @@
+module github.com/rshelekhov/golib/middleware/logging
+
+go 1.24.2
+
+require google.golang.org/grpc v1.74.2
+
+require (
+	go.opentelemetry.io/otel v1.37.0 // indirect
+	go.opentelemetry.io/otel/sdk/metric v1.37.0 // indirect
+	golang.org/x/net v0.42.0 // indirect
+	golang.org/x/sys v0.34.0 // indirect
+	golang.org/x/text v0.27.0 // indirect
+	google.golang.org/genproto/googleapis/rpc v0.0.0-20250721164621-a45f3dfb1074 // indirect
+	google.golang.org/protobuf v1.36.6 // indirect
+)
