+++ middleware/logging/http.go
@@ -1,0 +1,52 @@
+package logging
+
+import (
+	"log/slog"
+	"net/http"
+	"time"
+)
+
+// Middleware creates middleware for logging HTTP requests
+func Middleware(logger *slog.Logger) func(http.Handler) http.Handler {
+	return func(next http.Handler) http.Handler {
+		return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+			start := time.Now()
+
+			// Create a response wrapper to capture status
+			wrapper := &responseWriter{ResponseWriter: w, status: http.StatusOK}
+
+			// Process request
+			next.ServeHTTP(wrapper, r)
+
+			// Log the request
+			duration := time.Since(start)
+			logger.Info("http request",
+				"method", r.Method,
+				"path", r.URL.Path,
+				"status", wrapper.status,
+				"duration", duration,
+				"user_agent", r.UserAgent(),
+			)
+		})
+	}
+}
+
+// responseWriter is a wrapper for http.ResponseWriter that captures the status code
+type responseWriter struct {
+	http.ResponseWriter
+	status int
+}
+
+// WriteHeader captures the status code before delegating to the wrapped ResponseWriter
+func (rw *responseWriter) WriteHeader(code int) {
+	rw.status = code
+	rw.ResponseWriter.WriteHeader(code)
+}
+
+// Write captures a 200 status code if WriteHeader hasn't been called
+func (rw *responseWriter) Write(b []byte) (int, error) {
+	if rw.status == 0 {
+		rw.status = http.StatusOK
+	}
+	return rw.ResponseWriter.Write(b)
+}
