--- server/middleware.go
@@ -1,103 +1,0 @@
-package server
-
-import (
-	"log/slog"
-	"net/http"
-	"time"
-)
-
-// LoggingMiddleware creates middleware for logging HTTP requests
-func LoggingMiddleware(logger *slog.Logger) func(http.Handler) http.Handler {
-	return func(next http.Handler) http.Handler {
-		return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-			start := time.Now()
-
-			// Create a response wrapper to capture status
-			wrapper := &responseWriter{ResponseWriter: w, status: http.StatusOK}
-
-			// Process request
-			next.ServeHTTP(wrapper, r)
-
-			// Log the request
-			duration := time.Since(start)
-			logger.Info("http request",
-				"method", r.Method,
-				"path", r.URL.Path,
-				"status", wrapper.status,
-				"duration", duration,
-				"user_agent", r.UserAgent(),
-			)
-		})
-	}
-}
-
-// RecoveryMiddleware creates middleware for recovering from panics
-func RecoveryMiddleware(logger *slog.Logger) func(http.Handler) http.Handler {
-	return func(next http.Handler) http.Handler {
-		return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-			defer func() {
-				if err := recover(); err != nil {
-					// Log the error
-					logger.Error("http server panic recovered",
-						"error", err,
-						"path", r.URL.Path,
-						"method", r.Method,
-					)
-
-					// Return 500 error
-					http.Error(w, "Internal Server Error", http.StatusInternalServerError)
-				}
-			}()
-
-			next.ServeHTTP(w, r)
-		})
-	}
-}
-
-// CORSMiddleware creates middleware for handling CORS
-func CORSMiddleware(origins []string) func(http.Handler) http.Handler {
-	return func(next http.Handler) http.Handler {
-		return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-			// Set CORS headers
-			w.Header().Set("Access-Control-Allow-Headers", "Content-Type, Authorization")
-			w.Header().Set("Access-Control-Allow-Methods", "GET, POST, PUT, DELETE, OPTIONS")
-
-			// Check if the origin is allowed
-			origin := r.Header.Get("Origin")
-			for _, allowed := range origins {
-				if allowed == "*" || allowed == origin {
-					w.Header().Set("Access-Control-Allow-Origin", origin)
-					break
-				}
-			}
-
-			// Handle preflight requests
-			if r.Method == "OPTIONS" {
-				w.WriteHeader(http.StatusOK)
-				return
-			}
-
-			next.ServeHTTP(w, r)
-		})
-	}
-}
-
-// responseWriter is a wrapper for http.ResponseWriter that captures the status code
-type responseWriter struct {
-	http.ResponseWriter
-	status int
-}
-
-// WriteHeader captures the status code before delegating to the wrapped ResponseWriter
-func (rw *responseWriter) WriteHeader(code int) {
-	rw.status = code
-	rw.ResponseWriter.WriteHeader(code)
-}
-
-// Write captures a 200 status code if WriteHeader hasn't been called
-func (rw *responseWriter) Write(b []byte) (int, error) {
-	if rw.status == 0 {
-		rw.status = http.StatusOK
-	}
-	return rw.ResponseWriter.Write(b)
-}
